Added additional VIA functionality
